% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniparc2uniprotkb.R
\name{uniparc2uniprotkb}
\alias{uniparc2uniprotkb}
\title{Get UniprotKB identifier from Uniparc identifier}
\usage{
uniparc2uniprotkb(x, dryrun = FALSE, verbose = TRUE)
}
\arguments{
\item{x}{uniparc identifier}

\item{dryrun}{logical, do not execute the function}

\item{verbose}{logical for functin feedback}
}
\value{
UniProtKB identifier
}
\description{
Get UniprotKB identifier from Uniparc identifier
}
\examples{
uniparc2uniprotkb("WP_014719252.1", dryrun=TRUE)
}
