% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprot2name.R
\name{uniprot2name}
\alias{uniprot2name}
\title{Get Gene name from UniprotIDr}
\usage{
uniprot2name(x, dryrun = FALSE, verbose = TRUE)
}
\arguments{
\item{x}{UniprotID}

\item{dryrun}{logical, do not execute the function}

\item{verbose}{logical for function feedback}
}
\value{
A Gene Name
}
\description{
Get Gene name from UniprotIDr
}
\examples{
uniprotID2name("EKY21295", dryrun=TRUE)
}
