% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniprotID2Sequence.R
\name{uniprotID2Sequence}
\alias{uniprotID2Sequence}
\title{Get fasta sequence from UniprotKB identifier}
\usage{
uniprotID2Sequence(x, dryrun = FALSE, verbose = TRUE)
}
\arguments{
\item{x}{uniprot identifier}

\item{dryrun}{logical, do not execute the function}

\item{verbose}{logical for functin feedback}
}
\value{
A fasta sequence
}
\description{
Get fasta sequence from UniprotKB identifier
}
\examples{
uniprotID2Sequence("A8K9I2", dryrun=TRUE)
}
