% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicTools.uniprot.R
\docType{package}
\name{GenomicTools.uniprot-package}
\alias{GenomicTools.uniprot-package}
\title{R Package To Interact with UniProt Database
GenomicTools.uniprot is a loose collection of tools to interact with the UniProt database}
\description{
\tabular{ll}{ Package: \tab GenomicTools.uniprot\cr Type: \tab Package\cr Version: \tab
0.2\cr Date: \tab 2024-01-16\cr License: \tab GPL\cr LazyLoad: \tab yes\cr
}
}
\author{
Daniel Fischer

Maintainer: Daniel Fischer <daniel.fischer@luke.fi>
}
